Refactor dashboard into modular components
