+++ components/dashboard/hooks/use-auto-checker.ts
@@ -1,0 +1,1 @@
+export { useAutoCheck as useAutoChecker } from "@/hooks/use-auto-check";
